Remove Bezier class
